Implement delivery part reintegration and enhance solution evaluation
- Added a new method in DistributionOperations to reintegrate lost delivery parts into available vehicles based on capacity.
- Updated SolutionEvaluator to return a SolutionCost object instead of a double, improving cost evaluation granularity.
- Enhanced Solution class to encapsulate cost details more effectively.
- Refactored SolutionGenerator to handle scenarios where deliveries cannot be completed due to vehicle constraints.
- Improved unit tests to validate the new reintegration logic and ensure correct solution evaluation.

+++ src/main/java/com/example/plgsystem/pathfinding/PathFinder.java
@@ -40,8 +40,8 @@ public class PathFinder {
             closedSet.add(current.position);
 
             for (int[] dir : direcciones) {
-                int newX = current.position.getX() + dir[0];
-                int newY = current.position.getY() + dir[1];
+                double newX = current.position.getX() + dir[0];
+                double newY = current.position.getY() + dir[1];
 
                 if (newX < 0 || newX >= Constants.CITY_X ||
                         newY < 0 || newY >= Constants.CITY_Y) {
